requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
databases==0.6.1
fastapi==0.79.0
frozenlist==1.3.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
multidict==6.0.2
psycopg2==2.9.3
pydantic==1.9.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
typing-extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
yarl==1.8.1
